requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langsmith>=0.2.0
pydantic>=2.0
chromadb>=0.5.0
openai>=1.0.0           # xAI Grok API compatible with OpenAI SDK
tavily-python>=0.5.0
yfinance>=0.2.30
beautifulsoup4>=4.12.0
streamlit>=1.40.0
plotly>=5.20.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0